+++ PowerPointLabs/PowerPointLabs/PositionsLab/PositionsPaneWPF.xaml.cs
@@ -133,7 +133,7 @@ namespace PowerPointLabs.PositionsLab
 
             foreach (var currentShape in _shapesToBeRotated)
             {
-                PositionsLabMain.Rotate(currentShape, origin, angle);
+                PositionsLabMain.Rotate(currentShape, origin, angle, PositionsLabMain.ReorientShapeOrientation);
             }
         }
 
@@ -343,7 +343,7 @@ namespace PowerPointLabs.PositionsLab
 
             foreach (var currentShape in _shapesToBeRotated)
             {
-                PositionsLabMain.Rotate(currentShape, origin, angle);
+                PositionsLabMain.Rotate(currentShape, origin, angle, PositionsLabMain.ReorientShapeOrientation);
             }
 
             _prevMousePos = p;
